Renamed RequestType.ts to Request.ts
